--- src/canvas/Loader/Loader.tsx
@@ -1,30 +1,0 @@
-import { Html, useProgress } from "@react-three/drei"
-import styles from "./Loader.module.scss"
-
-export function CanvasLoader() {
-    const { progress } = useProgress()
-    return (
-        <Html
-            as="div"
-            center
-            style={{
-                display: "flex",
-                justifyContent: "center",
-                alignItems: "center",
-                flexDirection: "column",
-            }}
-        >
-            <span className={styles.loader}></span>
-            <p
-                style={{
-                    fontSize: 14,
-                    color: "#F1F1F1",
-                    fontWeight: 800,
-                    marginTop: 40,
-                }}
-            >
-                {progress.toFixed(2)}%
-            </p>
-        </Html>
-    )
-}
