--- src/canvas/StarsCanvas/StarsCanvas.module.scss
@@ -1,4 +1,0 @@
-.starsBg {
-    position: fixed;
-    z-index: -1;
-}
