--- src/canvas/index.ts
@@ -1,4 +1,0 @@
-import { ThreeCanvas } from "./DeathStarCanvas/ThreeCanvas"
-import { Loader } from "@react-three/drei"
-
-export { ThreeCanvas, Loader }
