+++ src/modules/Canvas/lib/getCurrentDevice.tsx
