+++ src/modules/Canvas/lib/math/basics.ts
@@ -1,0 +1,4 @@
+export const PI = 3.1415
+export const toRad = (n: number) => (n * PI) / 180
+export const sq = (n: number) => n ** 2
+export const { atan2, sqrt } = Math
