+++ src/modules/Canvas/types/types.ts
@@ -1,0 +1,1 @@
+export type Position = [number, number, number]
