+++ src/modules/Canvas/ui/Canvas.module.scss
