+++ src/modules/Canvas/ui/models/Stars/Stars.tsx
