+++ src/modules/Canvas/ui/models/Sun/Sun.tsx
@@ -8,7 +8,6 @@ export function Sun() {
 
     useFrame(() => {
         if (!sunRef.current) return
-
         sunRef.current.rotation.y += 0.000345
     })
